feat(TcpSocket): add data type check prevent assign exception (#548)
* test: 消除提示信息

* refactor: 使用扩展方法重构代码

* refactor: 增加赋值逻辑判断

* test: 更新单元测试

* chore: bump version 9.0.3

* chore: bump version 9.0.2

* refactor: 更正单词拼写错误

* refactor: 优化检查逻辑

* refactor: 更改绑定配置

* chore: 更新 Header 注释

+++ test/UnitTestTcpSocket/TcpSocketPropertyConverterTest.cs
@@ -67,6 +67,6 @@ public class TcpSocketPropertyConverterTest
     {
         var converter = new DataDoubleLittleEndianConverter();
         var actual = converter.Convert(new byte[] { 0x1F, 0x85, 0xEB, 0x51, 0xB8, 0x1E, 0x09, 0x40 });
-        Assert.Equal((double)3.14, actual);
+        Assert.Equal(3.14, actual);
     }
 }
